adjust profile record tab
